Rework: trust-only check, inline in the link workflows
Drop reachability probing entirely — many links are NIPRNet-only, so
clearnet probing proves nothing in either direction (even DNS failure
can mean internal-only). The check is now purely domain trust:
.mil/.gov/.edu or an allowlist entry.

Run inline in link_add, link_override, and link_update_af right before
commit, warning as a comment on the originating issue with instructions
for acknowledging a legitimate host in links_trust_allow.json. Bot
commits cannot trigger other workflows, so inline steps are also the
only reliable trigger point. The standalone workflow becomes a
dispatch-only full audit.

Allowlist entries now carry a reason each, documenting why every
commercial-domain service (software factories, GSA SmartPay, TRICARE
contractors, MWR) is considered official. The three .edu institutions
move out of the allowlist under blanket .edu trust.
