+++ .github/workflows/link_update_af.yml
@@ -49,6 +49,23 @@ jobs:
     - name: Overwrite links with new link list
       run: mv /tmp/links_af_new.json src/links/links_af.json
 
+    - name: Trust check
+      id: trust
+      run: node scripts/check-trust.js
+
+    - name: Warn on non-official domains
+      if: fromJSON(steps.trust.outputs.findings || '0') > 0
+      uses: actions/github-script@v9
+      with:
+        script: |
+          const fs = require('fs');
+          await github.rest.issues.createComment({
+            issue_number: context.payload.client_payload.original_issue,
+            owner: context.repo.owner,
+            repo: context.repo.repo,
+            body: fs.readFileSync('/tmp/trust_warning.md', 'utf8'),
+          });
+
     - name: Commit New Changes
       uses: EndBug/add-and-commit@v10
       with:
